extcon: intel-cht-wc: Add usb_id sysfs attribute
Some boards do not have the usb-id pin of the Whiskey Cove PMIC
hooked up to the connector, leaving the connector permanently in
gadget mode.

This commit adds a usb_id sys attribute which user can write to to
overwrite the usb-id pin value detected by the hardware so that
they can use their OTG connector in host or ACA mode despite the
usb-id pin not being connected.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>