power: supply: bq24190_charger: Make battery iface registraton conditional
The TI bq24190 chip is purely a charger chip, as such it has very
little battery info to export to userspace and in practice it is almost
always used together with a separate fuel-gauge chip which does have
interesting info to export.

Since the userspace ABI requires there to be only 1 power_supply device
registered per physical battery, the limited battery power_supply device
the bq24190_charger driver is registering is actually getting in the way
of a fuel-gauge driver exporting more useful info.

This commit adds a check for a "disable-battery-power-supply-iface"
device-property which can be set by platform-code if it knows that
there will be a fuel-gauge which will supply a battery interface.

Cc: Liam Breck <kernel@networkimprov.net>
Cc: Tony Lindgren <tony@atomide.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Changes in v2:
-This is a new patch in v2 of this patch-set
Changes in v3:
-Check for "disable-battery-power-supply-iface" device-property instead of
completely removing the battery interface